chore : 일단 랜딩페이지 리다이렉트 처리
