+++ apps/ticket/src/components/home/Home.tsx
@@ -1,7 +1,9 @@
-import { Button, Footer } from '@dudoong/ui';
+import { Button, ButtonSet, Footer, ListHeader } from '@dudoong/ui';
 import useGlobalOverlay from '@lib/hooks/useGlobalOverlay';
 import DDHead from '@components/shared/Layout/NextHead';
+import { useRouter } from 'next/router';
 const Home = () => {
+  const router = useRouter();
   const { openGlobalOverlay } = useGlobalOverlay();
   const openLoginTest = () => {
     openGlobalOverlay({ content: 'login' });
@@ -10,7 +12,11 @@ const Home = () => {
     <>
       <DDHead title="두둥! | 홈" />
       <main>
-        <Button onClick={openLoginTest}>로그인 열기</Button>
+        <ListHeader title={'테스트페이지'} size="listHeader_28" />
+        <ButtonSet varient="horizontal">
+          <Button onClick={openLoginTest}>로그인 열기</Button>
+          <Button onClick={() => router.push('/events/1')}>event1</Button>
+        </ButtonSet>
       </main>
       <section>
         <Footer />
