+++ apps/ticket/src/components/home/Landing.tsx
@@ -1,17 +1,13 @@
 import { Button, ButtonSet, Footer, ListHeader } from '@dudoong/ui';
 import DDHead from '@components/shared/Layout/NextHead';
 import { useRouter } from 'next/router';
+import { GetStaticProps } from 'next';
 const Landing = () => {
   const router = useRouter();
   return (
     <>
       <DDHead title="두둥!" />
-      <main>
-        <ListHeader title={'테스트페이지'} size="listHeader_28" />
-        <ButtonSet>
-          <Button onClick={() => router.push('/events/1')}>event1</Button>
-        </ButtonSet>
-      </main>
+      <main></main>
       <section>
         <Footer />
       </section>
@@ -20,3 +16,12 @@ const Landing = () => {
 };
 
 export default Landing;
+
+export const getStaticProps: GetStaticProps = () => {
+  return {
+    redirect: {
+      destination: '/home',
+      permanent: true,
+    },
+  };
+};
