+++ shared/ui/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.9",
+    "@dudoong/utils": "workspace:shared/utils",
     "emotion-reset": "^3.0.1",
     "polished": "^4.2.2",
     "react": "^18.2.0",
